Added left/right raycast to fish to better obstacle avoidance
Forward raycasting does not work properly to avoid some obstacles if the fish is too close and perpendicular to the obstacle. A turn due to the flock behaviour or following the flock target can get the origin of the raycast get within the mesh of the obstacle in the next frame. Two more small raycasts are added left and right to separate the fish from the mesh when it is close enough to fix that.